+++ tests/validator/test_validation_graph.py
@@ -13,33 +13,6 @@ from great_expectations.validator.validation_graph import (
 
 
 @pytest.fixture
-def table_head_metric_config() -> MetricConfiguration:
-    return MetricConfiguration(
-        metric_name="table.head",
-        metric_domain_kwargs={
-            "batch_id": "abc123",
-        },
-        metric_value_kwargs={
-            "n_rows": 5,
-        },
-    )
-
-
-@pytest.fixture
-def column_histogram_metric_config() -> MetricConfiguration:
-    return MetricConfiguration(
-        metric_name="column.histogram",
-        metric_domain_kwargs={
-            "batch_id": "def456",
-        },
-        metric_value_kwargs={
-            "bins": 5,
-        },
-        metric_dependencies=None,
-    )
-
-
-@pytest.fixture
 def metric_edge(
     table_head_metric_config: MetricConfiguration,
     column_histogram_metric_config: MetricConfiguration,
